Add buildQueryFromObjetc func
